base-service/requirements.txt
dj-database-url==0.5.0
Django==3.0.8
django-extensions==3.0.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-filter==2.3.0
